moved role enum
