+++ src/app/(user)/(users)/users/_components/AddUserDialog/AddUserDialog.tsx
@@ -11,7 +11,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog"
-import { Role } from "@/app/_lib/userConstants"
+import { Role } from "@/enums/role"
 import { UsersComboBox } from "./UsersComboBox"
 import { addUser } from "../../_lib/actions"
 import useSession from "@/hooks/useSession"
