+++ src/app/(user)/(users)/users/_lib/actions.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { Role } from "@/app/_lib/userConstants"
+import { Role } from "@/enums/role"
 import CanDoModel from "@/types/canDoModel"
 import { revalidateTag } from "next/cache"
