+++ src/app/(user)/_components/Header/RoleSelect.tsx
@@ -1,4 +1,4 @@
-import { Role } from "@/app/_lib/userConstants"
+import { Role } from "@/enums/role"
 import {
   Select,
   SelectContent,
